refactor: move src to new format, single entry build that builds easier
