+++ src/client/tsconfig.json
@@ -3,10 +3,12 @@
   "compilerOptions": {
     "lib": ["es2017", "dom"],
     "types": ["@citizenfx/client", "@types/node"],
+    "composite": true
   },
   "include": ["./**/*"],
-  "typedocOptions": {
-    "entryPoints": ["index.ts"],
-    "out": "../../docs"
-  }
+  "references": [
+    {
+      "path": "../common/tsconfig.json"
+    }
+  ]
 }
