+++ src/index.ts
@@ -1,4 +1,4 @@
 export * from "./client";
 export * from "./server";
-export * from "./misc";
-export * from "./types";
+export * from "./common/helpers";
+export * from "./common/types";
