--- src/misc/constants.ts
@@ -1,12 +1,0 @@
-export enum ServerErrorCodes {
-  UnknownError = "UNKNOWN_ERROR",
-  TimedOut = "TIMED_OUT",
-  NotAllowed = "NOT_ALLOWED",
-  ServerError = "SERVER_ERROR",
-  InvalidData = "INVALID_DATA",
-}
-
-export enum ServerStatus {
-  OK = "ok",
-  Error = "error",
-}
