+++ src/server/tsconfig.json
@@ -1,11 +1,13 @@
 {
   "extends": "../../tsconfig.json",
   "compilerOptions": {
-    "types": ["@citizenfx/server", "@types/node"]
+    "types": ["@citizenfx/server", "@types/node"],
+    "composite": true
   },
   "include": ["./**/*"],
-  "typedocOptions": {
-    "entryPoints": ["index.ts"],
-    "out": "../../docs"
-  }
+  "references": [
+    {
+      "path": "../common/tsconfig.json"
+    }
+  ]
 }
