+++ src/tsconfig.json
@@ -1,0 +1,18 @@
+{
+  "extends": "../tsconfig.json",
+  "include": ["index.ts"],
+  "typedocOptions": {
+    "entryPoints": ["./server/index.ts", "./client/index.ts", "./common/helpers.ts", "./common/types.ts"]
+  },
+  "references": [
+    {
+      "path": "./client/tsconfig.json"
+    },
+    {
+      "path": "./server/tsconfig.json"
+    },
+    {
+      "path": "./common/tsconfig.json"
+    }
+  ]
+}
